Made all filenames lowercase (Sorry if this will upset anyone)
