+++ ArduCAM4Pi/Makefile
@@ -1,44 +1,44 @@
-all : OV2640_Playback	OV3640_Playback	OV5642_Playback	OV7670_Playback OV2640_DigitalCamera	OV5642_DigitalCamera OV2640_Capture
+all : ov2640_playback	ov3640_playback	ov5642_playback	ov7670_playback ov2640_digital_camera	ov5642_digital_camera ov2640_capture
 
-objects = DefaultFonts_SPI.o UTFT_SPI.o PiCAM.o
+objects = default_fonts_spi.o utft_spi.o picam.o
 
-OV2640_Playback : $(objects) PiCAM_OV2640_Playback.o 
-	gcc -o OV2640_Playback $(objects) PiCAM_OV2640_Playback.o -lwiringPi -Wall
-OV3640_Playback : $(objects) PiCAM_OV3640_Playback.o 
-	gcc -o OV3640_Playback $(objects) PiCAM_OV3640_Playback.o -lwiringPi -Wall
-OV5642_Playback : $(objects) PiCAM_OV5642_Playback.o 
-	gcc -o OV5642_Playback $(objects) PiCAM_OV5642_Playback.o -lwiringPi -Wall
-OV7670_Playback : $(objects) PiCAM_OV7670_Playback.o 
-	gcc -o OV7670_Playback $(objects) PiCAM_OV7670_Playback.o -lwiringPi -Wall
-OV2640_DigitalCamera : $(objects) PiCAM_OV2640_DigitalCamera.o 
-	gcc -o OV2640_DigitalCamera $(objects) PiCAM_OV2640_DigitalCamera.o -lwiringPi -Wall
-OV2640_Capture : $(objects) PiCAM_OV2640_Capture.o 
-	gcc -o OV2640_Capture $(objects) PiCAM_OV2640_Capture.o -lwiringPi -Wall
-OV5642_DigitalCamera : $(objects) PiCAM_OV5642_DigitalCamera.o 
-	gcc -o OV5642_DigitalCamera $(objects) PiCAM_OV5642_DigitalCamera.o -lwiringPi -Wall
+ov2640_playback : $(objects) picam_ov2640_playback.o 
+	gcc -o ov2640_playback $(objects) picam_ov2640_playback.o -lwiringPi -Wall
+ov3640_playback : $(objects) picam_ov3640_playback.o 
+	gcc -o ov3640_playback $(objects) picam_ov3640_playback.o -lwiringPi -Wall
+ov5642_playback : $(objects) picam_ov5642_playback.o 
+	gcc -o ov5642_playback $(objects) picam_ov5642_playback.o -lwiringPi -Wall
+ov7670_playback : $(objects) picam_ov7670_playback.o 
+	gcc -o ov7670_playback $(objects) picam_ov7670_playback.o -lwiringPi -Wall
+ov2640_digital_camera : $(objects) picam_ov2640_digital_camera.o 
+	gcc -o ov2640_digital_camera $(objects) picam_ov2640_digital_camera.o -lwiringPi -Wall
+ov2640_capture : $(objects) picam_ov2640_capture.o 
+	gcc -o ov2640_capture $(objects) picam_ov2640_capture.o -lwiringPi -Wall
+ov5642_digital_camera : $(objects) picam_ov5642_digital_camera.o 
+	gcc -o ov5642_digital_camera $(objects) picam_ov5642_digital_camera.o -lwiringPi -Wall
 	
-DefaultFonts_SPI.o : DefaultFonts_SPI.c
-	gcc -c DefaultFonts_SPI.c -lwiringPi -Wall
-UTFT_SPI.o : UTFT_SPI.c
-	gcc -c UTFT_SPI.c -lwiringPi -Wall
-PiCAM.o : PiCAM.c
-	gcc -c PiCAM.c -lwiringPi -Wall
-PiCAM_OV2640_Playback.o : PiCAM_OV2640_Playback.c
-	gcc -c PiCAM_OV2640_Playback.c -lwiringPi -Wall
-PiCAM_OV3640_Playback.o : PiCAM_OV3640_Playback.c
-	gcc -c PiCAM_OV3640_Playback.c -lwiringPi -Wall
-PiCAM_OV5642_Playback.o : PiCAM_OV5642_Playback.c
-	gcc -c PiCAM_OV5642_Playback.c -lwiringPi -Wall
-PiCAM_OV7670_Playback.o : PiCAM_OV7670_Playback.c
-	gcc -c PiCAM_OV7670_Playback.c -lwiringPi -Wall
-PiCAM_OV2640_DigitalCamera.o : PiCAM_OV2640_DigitalCamera.c
-	gcc -c PiCAM_OV2640_DigitalCamera.c -lwiringPi -Wall
-PiCAM_OV2640_Capture.o : PiCAM_OV2640_Capture.c
-	gcc -c PiCAM_OV2640_Capture.c -lwiringPi -Wall
-PiCAM_OV5642_DigitalCamera.o : PiCAM_OV5642_DigitalCamera.c
-	gcc -c PiCAM_OV5642_DigitalCamera.c -lwiringPi -Wall	
+default_fonts_spi.o : default_fonts_spi.c
+	gcc -c default_fonts_spi.c -lwiringPi -Wall
+utft_spi.o : utft_spi.c
+	gcc -c utft_spi.c -lwiringPi -Wall
+picam.o : picam.c
+	gcc -c picam.c -lwiringPi -Wall
+picam_ov2640_playback.o : picam_ov2640_playback.c
+	gcc -c picam_ov2640_playback.c -lwiringPi -Wall
+picam_ov3640_playback.o : picam_ov3640_playback.c
+	gcc -c picam_ov3640_playback.c -lwiringPi -Wall
+picam_ov5642_playback.o : picam_ov5642_playback.c
+	gcc -c picam_ov5642_playback.c -lwiringPi -Wall
+picam_ov7670_playback.o : picam_ov7670_playback.c
+	gcc -c picam_ov7670_playback.c -lwiringPi -Wall
+picam_ov2640_digital_camera.o : picam_ov2640_digital_camera.c
+	gcc -c picam_ov2640_digital_camera.c -lwiringPi -Wall
+picam_ov2640_capture.o : picam_ov2640_capture.c
+	gcc -c picam_ov2640_capture.c -lwiringPi -Wall
+picam_ov5642_digital_camera.o : picam_ov5642_digital_camera.c
+	gcc -c picam_ov5642_digital_camera.c -lwiringPi -Wall	
 
 clean : 
-	rm -f OV2640_Playback OV3640_Playback OV5642_Playback OV7670_Playback OV2640_DigitalCamera OV2640_Capture OV5642_DigitalCamera $(objects) \
-	PiCAM_OV2640_Playback.o PiCAM_OV3640_Playback.o PiCAM_OV5642_Playback.o PiCAM_OV7670_Playback.o \
-	PiCAM_OV2640_DigitalCamera.o PiCAM_OV5642_DigitalCamera.o
+	rm -f ov2640_playback ov3640_playback ov5642_playback ov7670_playback ov2640_digital_camera ov2640_capture ov5642_digital_camera $(objects) \
+	picam_ov2640_playback.o picam_ov3640_playback.o picam_ov5642_playback.o picam_ov7670_Playback.o \
+	picam_ov2640_digital_camera.o picam_ov5642_digital_camera.o
