+++ ArduCAM4Pi/ov5642_regs.h
@@ -1,6 +1,6 @@
 #ifndef OV5642_REGS_H
 #define OV5642_REGS_H
-#include "PiCAM.h"
+#include "picam.h"
 
 #define OV5642_CHIPID_HIGH 0x300a
 #define OV5642_CHIPID_LOW 0x300b
