+++ ArduCAM4Pi/picam_ov2640_capture.c
@@ -30,7 +30,7 @@
 #include <time.h>
 #include <wiringPiSPI.h>
 #include <unistd.h>
-#include "PiCAM.h"
+#include "picam.h"
 
 #define OV2640_CHIPID_HIGH  0x0A
 #define OV2640_CHIPID_LOW   0x0B
